Achieve PHPStan max-level compliance with zero errors
Install Larastan v3.9.2 for Laravel-aware static analysis and fix all
364 type errors across models, services, Livewire components, jobs,
middleware, and providers. Add @property PHPDoc annotations to 16+
Eloquent models, type assertions throughout services, and generic
type parameters for collections and relationships.

+++ app/Services/TaxCalculator.php
@@ -12,8 +12,9 @@ class TaxCalculator
      */
     public function calculate(int $amount, TaxSettings $settings, array $address): TaxLine
     {
+        /** @var array{default_rate?: int} $config */
         $config = $settings->config_json ?? [];
-        $rate = (int) ($config['default_rate'] ?? 0);
+        $rate = $config['default_rate'] ?? 0;
 
         if ($rate === 0 || $amount === 0) {
             return new TaxLine(
